Rename options package from xlator/options to options
