+++ glusterd2/xlator/load.go
@@ -16,7 +16,7 @@ import (
 	"strings"
 	"unsafe"
 
-	"github.com/gluster/glusterd2/glusterd2/xlator/options"
+	"github.com/gluster/glusterd2/glusterd2/options"
 	"github.com/gluster/glusterd2/pkg/utils"
 
 	log "github.com/sirupsen/logrus"
