+++ glusterd2/xlator/xlator.go
@@ -1,7 +1,7 @@
 package xlator
 
 import (
-	"github.com/gluster/glusterd2/glusterd2/xlator/options"
+	"github.com/gluster/glusterd2/glusterd2/options"
 )
 
 // Xlator represents a GlusterFS xlator
